requirements.txt
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
